requirements.txt
fastapi
uvicorn
jinja2
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
sqlalchemy
psycopg2-binary
alembic

python-multipart

grpcio
grpcio-tools